refactor ai21_base + segmentation
